requirements.txt
PyGithub==2.1.1
anthropic>=0.18.0
slack-sdk==3.26.1
python-dotenv==1.0.0
